Fix more missing renames (#2)
This fixes a number of missed renames in the initial push.

* Rename a2/lib to automate/lib

Our Go-based services use build-time linker flags to inject version
information into the resulting executable. In our initial rename we
missed the paths to these linker flags. This PR changes `lib/a2` to
`lib/automate` in the relevant plans.

* Allow building even when the dev manifest is wrong
* Rebuild protobufs
* Fix a2 references in concurrency group config

Signed-off-by: Steven Danna <steve@chef.io>