[DEPLOY-513] Add backup cancellation
During backup and restore there are lots of I/O operations which we've
previously defined hardcoded timeouts for. Even worse, some of the
underlying libraries that do I/O don't respect timeouts at all. In this
change we update backup code to use a singular timeout for operations
that support timeouts or retries.

A drawback to this approach is that with the default timeout of 2 hours,
we might end up in failure loops where we are waiting for until the
timeout expires to fail.

To handle these cases we've created a singular backup.Runner instance on
the deployment-service's server instance. All backup operations now
share that singular instance and we've added cancellation support for
the currently running operation.

Two `chef-automate` backup commands have been added to interface with
this new functionality:

* `chef-automate backup status` returns the status of the backup runner.
  If the runner is idle it will return that as the state.

* `chef-automate backup cancel` will cancel the currently running
  operation. Create, Delete, and Restore are the only cancellable
  operations as they are the only operations that change state. If the
  command is run when no operations are running it will return an error.

We also introduce per-command wait timeouts to ensure commands don't
stay in retry loops for longer than necessary before the exit.

BUGFIX:

Fix event stream deadlock

It was possible for event stream to deadlock, which this patch fixes.
StreamTo is no longer responsible for maintaining a channel. Instead,
functions are registered. If it returns an error, it is deregistered
by the event loop, and not done from another goroutine.

More info starts here:
https://github.com/chef/a2/pull/5424#issuecomment-480964963

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>