Remove context based lock acquisition
This code wasn't correct and could not be correct. If the mutex is
locked after the context expires, there's no way to unlock it. So if you
block trying to take the mutex, you must unlock it.

To get around this, we're going to check the context for expiration
before trying to take the lock. If it's expired, we wont try to take it.
Otherwise, we'll take the mutex. When the context is next checked, it
will return an expiration error and we'll know that we need to unlock
the mutex

The other option is to replace the standard libary mutex with something
that uses channels as a mutex.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>