(wip) gateway: refactor authorizer
this is close but not quite there.

the problem is how to pass the right minor version in the gateway handler.

and how to differentiate btwn v2 and v2.1 request.

might need to think harder about the approach of using a new authorizer
entirely that's just for switching btwn v2 and v1.

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>