create-manifest.rb: pull packages from dev channel
Using the unstable channel was working because the only process that
uploaded packages to our origin was our build process. At the moment,
release engineering is testing a new build process and as a result,
unstable has unexpected packages.

This defends us against that situation by only pulling package from
the dev channel.

Signed-off-by: Steven Danna <steve@chef.io>