[workflow-ctl] Remove more unused commands
This change removes the following commands and their dependencies from
workflow-ctl:

```
cleanse
data-summary
delete-node
doctor
graceful-kill
hup
int
kill
node-summary
once
reconfigure
restart
service-list
show-config
start
status
stop
tail
term
uninstall
usr1
usr2
```

As part of those removals it:

- Replaces `omnibus-ctl` with a small stub that has all of the
  functionality we were actually using from omnibus-ctl.

- Removes `lib/setup` which was used by the already removed `setup`
  command

- Renames most uses of `automate-ctl` to `workflow-ctl` inside the code

One of the riskier changes here is that the previous code was assuming
omnibus-ctl style argument ordering (where the first couple arguments
are omnibus-ctl config) whereas now we assume that we care about all
arguments. This is risky because many of the remaining commands still
do not work and thus it is hard to test that the modifications are
correct.

Signed-off-by: Steven Danna <steve@chef.io>