WIP [workflow-nginx] Remove maintenance mode and mime.types
Chef Automate has a maintenance mode implemented in
automate-load-balancer. As far as I can tell there was no real way to
turn on this maintenance mode anymore, so I've removed it.

Also, I've removed the vendored mime.types as it was identical to the
one in the openresty package we already depend on.

Signed-off-by: Steven Danna <steve@chef.io>