jobs list and node read fixes (#56)
* only show report button on scan job list if not recurring job

a scan job with a recurrence schedule should not show a report
button on the jobs list. the report button should only show
up on the actual scans list for recurring jobs.

* do not fail node read if secret does not exist

if we cannot find the credential id attached to the node,
we should not fail the node. we should just return what is
found and leave it to the user from there. failing the node
results in unreadable nodes, which helps no one.

Signed-off-by: Victoria Jeffrey <vjeffrey@chef.io>

* Update components/nodemanager-service/api/nodes/server/server.go

Co-Authored-By: vjeffrey <vjeffrey@chef.io>
Signed-off-by: Victoria Jeffrey <vjeffrey@chef.io>