(wip) authz: have authz server check version
we decided to try keeping the v2.0 vs. v2.1 decision internal to
authz-service. So now we call ProjectsAuthorized in the v2 middleware on
every request. ProjectAuthorized checks what version policy server's
version channel is set to, and overrides the projectFilter list if set
to v0.

currently the returning of the channel values in GetVersion needs work.

by using the channel instead of the GetPolicyVersion API call
we avoid the overhead of an extra API call.

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>