[integration] Generate config for deep_upgrade test using old CLI (#87)
Previously we were using the CLI from the current channel to generate
our configuration. This worked previously because the configuration
generation has changed very little over time. As of the most recent
promotion, however the deep_upgrade tests have been failing with:

```
Failure: Config file must be a valid automate config: unknown configuration keys:
    global.v1.frontend_tls
    global.v1.frontend_tls.cert
    global.v1.frontend_tls.key
Cause: unknown configuration keys:
        global.v1.frontend_tls
        global.v1.frontend_tls.cert
        global.v1.frontend_tls.key
```

because the old CLI version we use for the deploy does not understand
the new keys in the configuration generated by the CLI that was just
promoted to current.

While we typically avoid this kind of strict checking in the API, in
the CLI and configuration it helps prevent user typos from causing
confusion.

The solution here is to generate the configuration with the older
version of the automate-cli that we are using for the deep upgrade
test.

Signed-off-by: Steven Danna <steve@chef.io>