[deployment] Represent waiting states in converger
* Motivation

This PR started as a fix for the SupervisorUpgrade phase. That phase
was not restarting our systemd unit when required. Fixing that problem
revealed a deadlock in our systemd restart sequence. Namely,

- The Converge function would call `systemctl restart chef-automate`

- `systemctl restart chef-automate` would call the Stop() grpc call

- The Stop grpc call requires both the deployment lock (which the
  periodic converger likely held) and a responsive converger (which is
  unlikely as it is blocked waiting for `systemctl restart` to
  finish).

We solved the most immediate issue by passing the `--no-block` flag to
`systemctl restart` which allowed the converger/compiler to move on,
releasing the lock and responding to our Stop() request.

However, we were now faced with one of our common fears:

> How safe is it for us to be restarted in the middle of the compiler?

Previously we've solve this via three strategies:

- Making the compiler stop if it detected that a restart or
  reconfigure should be coming,

- Fixing bugs in `hab` that made killing commands unsafe, and

- Ad-hoc state mechanisms to prevent the compiler from moving forward.

While we could have added a new ad-hoc mechanism here, we decided we
wanted to model our waiting states in a small state machine.

* Converger State Machine

During upgrades and reconfiguration, we sometimes expect an external
process (hab-sup or systemd) to send our process a signal that will
result in our processes restarting. When this occurs, we want to wait
for that signal to arrive rather than starting new work that could be
interrupted.

To do that, the converger is now a small state machine that looks like
the following:

```
        +-----------+  convergeRequest           +-------------------+
        |           |  err = ErrRestartPending   |                   |
        |           +--------------------------->|                   |
        |    idle   |                            | waitingForRestart |
        |           |<---------------------------+                   |
        |           |       timeout              |                   |
        +-------+---+                            +-------------------+
            ^   |
            |   | convergeRequest
   timeout 2|   | err = ErrSelfReconfigurePending
            |   |
            |   v
         +--+--------------------+
         |                       | timeout 1
         |                       +---------+
         | waitingForReconfigure |         |
         |                       |<--------+
         |                       |
         +-----------------------+
```

Note that this state machine is expressing what I believe were the
previous semantics of the code.

* Converger version removal

Previously the Converge() request passed along a version. The idea of
this version was to make sure we weren't making changed based on
outdated deployment configuration.  I think this is a great idea but:

1. We weren't using it anywhere, and
2. For it to work we would have to centralize mutations to the
   deployment struct.

Thus, I've removed it for now.

* Debug Interface

The converger now has a debug channel where test code can get and set
the current state of the state machine. An alternative to this
approach would be to use a small mutex around the converger state.

* Follow-up Work

- Backup and Restore should probably be wired into this, however we
  want a week where we don't change the backup/restore code so we are
  deferring that work.

- Re-add the converge version more systematically.

- We really want to rename `converger` to `SOMETHING` and `compiler`
  to `converger` but I didn't do that here.

Signed-off-by: Steven Danna <steve@chef.io>