Add chef-automate subcommands to toggle applications tab features (#107)
Visibility into applications (habitat) data is currently alpha quality
and feature-flagged off by default. For architectural reasons, three
different feature toggles must be flipped to fully enable the feature on
the back-end. To make it easier for developers (and eventually users) to
test, this patch adds a CLI subcommand to toggle the feature. The
subcommand is currently hidden; we likely will unhide it in the future
to facilitate an open beta.

Signed-off-by: Daniel DeLeo <dan@chef.io>