Dan/applications load gen (#85)
Adds a load generator for testing the performance of applications service data ingestion. I designed the configuration of the load generator such that we can configure it for a variety of scenarios that emulate various customers' configurations without too much repetition. For a simple traffic generation scenario, I have included a basic built-in configuration that can be activated via command line flag.

* Create config for load generator w/ a command to explain it
* Implement load generation for apps svc
* Add statistics to apps svc load generator
* cleanup some TODOs
* Add a simple built-in load profile for quick testing
* Build `applications-load-gen` with the hab package
* Add some basic tests for load generator
* update go pkg lock
* Update expeditor config
* fix nits