Update golangci-lint to make it more usable in the hab studio
golangci-lint 1.15 is affected by a bug that occurs when in a path with
a symlink, which prevents it from properly applying `nolint` directives
in comments. In the studio, we symlink the automate source from /src
into the scaffolding gopath when you run `setup_go_workspace`; the
combination of these factors makes golangci-lint not usable in the
studio. 1.16 is supposed to contain a fix.

Signed-off-by: Daniel DeLeo <dan@chef.io>