Restore for external PG  (#82)
Make restore work with external PG. This introduces a new database backup operation and deprecates the old one, meaning for external PG backup/restore to work, you'll need to use the latest automate. The reason we need a new op is because we now use the custom pg dump format. This allows us to ignore restoring permissions, which is important because we no longer get to always control the role to use. These will get fixed up when services start up and ask pg-helper to do things.