Add a test to ensure load generator code stays updated
If we are not actively doing load test work, we are probably gonna
forget to update the load generator when we add some new data field to
our habitat events. Add a test that can detect when a new field is added
to the applications.HabService message and we don't add code to the load
generator to fill it in.

Signed-off-by: Daniel DeLeo <dan@chef.io>