Apply golangci-lint to config-mgmt-service and secrets-service
These are 2 of the 3 services that don't seem to be using the linter
currently. This applies the linter to those projects which will
hopefully make it a bit easier to transition to having
`go_component_static_tests` run the linter.

`ingest-service` is, I believe, likely the last service that needs to
have the linter applied to it. I have not done it here because there
are a large number of errcheck violations that need to be looked at
carefully.

Signed-off-by: Steven Danna <steve@chef.io>