(wip) authz: update ProjectsAuthorized
we have one more case we need to handle:
when the request contains no projects,
one statement allows all projects,
another statement denies at least one individual project.

we want this to return all projects except for the denied projects.

but the rego needs to know all the projects it needs to compare in order to
return the response above. that's not provided when the request contains
no projects.

so we use FilterAuthorizedProjects to get the list of all allowed
projects to that ProjectsAuthorized must check against.