Add TLS disable option to event-gw internal config
It would certainly be great if customers never ever use this, but
currently the rust NATS streaming library we want to use in hab doesn't
correctly support TLS the kinda weird way that NATS does it. While
waiting for the library to be updated (currently in progress), we want
to test our other behaviors end-to-end.

Signed-off-by: Daniel DeLeo <dan@chef.io>