authz: just call listProjects
previously, ProjectsAuthorized called FilterAuthorizedProjects
if the project filter was empty. this way it could provide the engine ProjectsAuthorized
with a list of projects, allowing it to account for cases when deny overrides
allowed projects

however, that's 2 opa calls in one handler, which is a lot of expensive opa work.

all we really need in ProjectsAuthorized is a complete list of projects,
so we can just call ListProjects and get what we need.
it is still an additional call,
but at least while we have a small max number of projects,
it's less costly than FilterAuthorizedProjects.

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>