Pin HA Backend packages
This is temporary, we'll fix it on monday. To get back to unpinned,
we need to update the configuration enable SSL for elasticsearch with
the ssl certs that we have generated, and tell automate about those
certs. Also, we need to update to hab > 0.75, or figure out why our
containers hab is so old.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>