[secrets-service] Use platform secrets store for shared secret
Currently, secrets-service stores the symmetric encryption key on
disk. Now, it uses the platform secrets code.

Right now the platform secrets store is also an insecure on-disk.
However, this makes it a bit easier for us to make sure we can export
all of the known shared secrets in the system.

To mitigate the very confusing naming, I've moved lib/secrets to
lib/platform/secrets.

Signed-off-by: Steven Danna <steve@chef.io>