[deployment] Persist the backup manifest on restore
Previously, we were not persisting the manifest at the end of the
restore process. This meant that the first converge after an "upgrade
restore" would potentially result in a downgrade if the user had their
upgrade strategy set to none because the converge loop would use the
old manifest.

TODO:

- [ ] Needs a test somewhere

Signed-off-by: Steven Danna <steve@chef.io>