[e2e] ensure we delete scan jobs after tests (#227)
in our long-lived test environments we're currently failing to delete
scan jobs for some reason. this change adds a test with the intent to
check that the scan job has been deleted before we proceed. this test
may end up being flaky, but we should fix the flakiness and ensure
that we're correctly deleting things in the test

Signed-off-by: Stephen Delano <stephen@chef.io>