Fix all paginators from the Applications Tab (#218)
We had a number of issues with our paginators, this change is fixing them all so that the UI and Backend work as expected. Now you can use all Paginators in the Applications Tab. That includes when using the health filters! 🎉 

Activities:
* Adds test to verify paginator inputs
* Fix Apps service-groups pagination in the UI

This code enables the pagination in the UI for service groups correctly,
when the user selects a Health Status Filter, we adjust the total number
of service groups to diplay in the page footer. Before we were always
displaying the Total of all service-groups.

* Fix Apps service-groups pagination in the Backend
* Rename test functions to match RPC func being tested
* Fix pagination in UI Apps services-sidebar

Signed-off-by: Salim Afiune <afiune@chef.io>