Materialize project filter on all pages
The filter is not displayed until  the IAM version is fetched.
Previously that was only done when routing to the Settings page.
Now it occurs:
(a) when loading the app
(b) periodically on an on-going basis
The latter will act on an IAM version change every 5 minutes
even if the user does not refresh the page.

Signed-off-by: michael sorens <msorens@users.noreply.github.com>