Project update resume (#132)
The project update process can be a long running process. Therefore if the compliance-service, ingest-service, or authz-service is updated during this process it will need to resume where it left off. To perform a resume the service needs to save its state during the process. Then during startup, the services need to check this saved state to resume the update process if needed.

A restart of the elasticsearch service does not allow the project update process to resume. If this service restarts the compliance-service and ingest-service will detect that their jobs are not running any longer and send a failure message to the authz-service. The authz-service will mark the update process as a failure with a failure message.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>