WIP SPIKE: workflow library
*This code is in flux and not ready for code-level review. We've opened
this PR to facilitate a discussion about the high-level design.*

This is a spike implementation of a "workflow" library. The goal of
this library is to provide applications with a means of describing
application-specific processes such that the can be run across
multiple machines and resumed after crashes or reboots.

This library is inspired by the following systems:

https://cadenceworkflow.io/
https://docs.aws.amazon.com/swf/index.html

It is centered around workflows (nothing to do with delivery) and
tasks. Workflows coordinate the execution of Tasks. Tasks are discrete
bits of application logic.

A Workflow can

- start Tasks,
- make decisions based on the completion of Task, and
- make decisions based on a cancellation signal.

The logic of a workflow is defined in a WorkflowExecutor: Go code that
application authors write and register with the library. The executor
is defined as a set of callbacks. The workflow executor is started
with user-defined parameters that allow the same workflow executor to
be reused. The workflow executor can store state in a payload that is
persistent between steps, allowing it to implement more complex logic
based on past events.

A simple WorkflowExecutor might look like this (API not final):

```golang
type SimpleWorkflowExecutor struct{}
type SimpleWorkflowPayload struct {
	State string
}

// The OnStart callback is the entry point for a workflow’s execution
func (*SimpleWorkflowExecutor) OnStart(w workflow.WorkflowInstanceHandler,
	ev workflow.StartEvent) workflow.Decision {

	// Request "simple_task" be run
	w.EnqueueTask("simple_task", SimpleTaskParams{EchoMe: "Weeeee"})
	return w.Continue(SimpleWorkflowPayload{State: "REMEMBER ME"})
}

// OnTaskComplete is called whenever a task related to this workflow completes
func (p *SimpleWorkflowExecutor) OnTaskComplete(w workflow.WorkflowInstanceHandler,
	ev workflow.TaskCompleteEvent) workflow.Decision {
	switch ev.TaskName {
	case "simple_task":
		if err := ev.Result.Err(); err != nil {
			logrus.WithError(err).Error("simple_task failed")
		} else {
			var result string
			ev.Result.Get(&result)
			logrus.WithField("result", result).Info("simple_task done")
		}
	default:
		logrus.Error("Unknown task")
	}
	return w.Complete()
}
// OnCancel is called if any user-code sends a cancel single to this workflow
func (p *SimpleWorkflowExecutor) OnCancel(w workflow.WorkflowInstanceHandler, ev workflow.CancelEvent) workflow.Decision {
	return w.Complete()
}

The logic of a task is defined in a Task Executor: Go code that
application authors write and register with the library. For example,
a simple TaskExecutor might look like this:

```golang
type SimpleTaskParams struct {
	EchoMe string
}

type SimpleTaskExecutor struct{}

func (t *SimpleTaskExecutor) Run(ctx context.Context, task workflow.TaskQuerier) (interface{}, error) {
	params := SimpleTaskParams{}
	if err := task.GetParameters(&params); err != nil {
		return nil, err
	}
	fmt.Printf("%s\n", params.EchoMe)
	return "echoed", nil
}
```

While structuring application logic in this way is not always
straight-forward, it comes with a benefit. As long as
WorkflowExecutors and TaskExecutors use the provided payload and
parameter functions to receive and store state, the execution of a
workflow can occur across multiple machines in a Chef Automate
installation and resume after an application crash or restart.

We’ve implemented this system as a library that can be included in any
Go-based service.  It is built on top of PostgreSQL. The
implementation is still in development and the current code can be
found in the following pull request.

TaskExecutors work by starting a transaction, selecting an unprocessed
task from the tasks table, running the user-level code related to that
task, and then committing a transaction that removes the unprocessed
task from the task list and adds the result to the task_result
table. Row-level locking via SELECT FOR UPDATE is used to ensure only
one executor runs a given task at a time. If a TaskExecutor
disappears, the postgresql session will end, aborting the transaction
and leaving the task for another executor to restart.  Note that this
implies that Tasks should generally strive to be re-runnable.

WorkflowExecutors operate similarly to TaskExecutors.

Scheduled workflows are built on a polling system that checks all
scheduled workflows and starts any workflows that are due to start.
Uniqueness constraints on the workflow_instances table help us
identify if the task is still running from the last invocation.

TODO:

- [ ] Tests
- [ ] Public-facing APIs for workflow cancellation
- [ ] Public-facing APIs for listing scheduled jobs
- [ ] Tests
- [ ] Tests
- [ ] Refactor type names. ssd has a plan

Co-authored-by: Jay Mundrawala <jmundrawala@chef.io>
Co-authored-by: Yvonne Lam <yvonne@chef.io>
Signed-off-by: Steven Danna <steve@chef.io>
Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>