Adding project filters for GetRunsCounts (#300)
Adding project filtering for the cfgmgmt.proto GetRunsCounts RPC.

Limited the functionality of the GetRunsCounts RPC. Now an API user has to provide the node ID to use for the counts. A user can not get the run counts of more than one node. This does not change the behavior of the UI.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>