[terraform] attempt to stabilize sysctl templates
recently we've seen a fair share of failures from various fresh
install nodes in our test environments that aren't finding the
a2-sysctl.conf file in /tmp or even after applying them aren't
seeing the results in the system config when running the automate
preflight checks.

this change attempts to stabilize the writing of the chef automate
sysctl config by using the terraform file `content` feature rather
than loading the template from a local file. we've not seen similar
failures recently with the `content` feature, so this eliminates
one potential point of failure.

additionally, the file we write to /etc/sysctl.d has been renamed
from `a2-sysctl.conf` to `60-chef-automate.conf` to prevent any
ambiguity in the sorting or application of the sysctl config files.

Signed-off-by: Stephen Delano <stephen@chef.io>