Fixing flaky test (#343)
Fixing a flaky UI e2e test. When selecting the event feed dropdown filter the 'active' attribute should be added. The problem was that the test was checking before that 'active' attribute was added. So, the fix here is to wait until after the click and then check for the 'active' attribute for about a second. The other option was to add a sleep statement and sleeps cause other problems.

I found a solution thanks to this website https://www.captechconsulting.com/blogs/keep-your-promises-making-the-jump-from-selenium-to-protractor
Signed-off-by: Lance Finfrock <lfinfrock@chef.io>