opa: benchmark fixes
The existing initPartialResultV2 benchmark did almost nothing: the store
that was set up was the v1 store, not the v2 store. This is fixed now.

Also, on a hunch, I've explored if we might have state issues: and
indeed, running the initPartialResultV2 function _on an unchanged store_
*multiple times* makes performance blow up.  I take this as an
indication that our problem is not the way our rego policies are laid
out, but rather some state business in our engine code.

Signed-off-by: Stephan Renatus <srenatus@chef.io>