automate-gateway: cleanup GRPC/HTTP multiplexing cruft
This was carried over from some ancient grpc-gateway example code[1]; and
we're not exposing the GRPC API that way right now.

Using (*grpc.Server).ServeHTTP has some issues:

1. it's known to be slow, very slow[2]
2. it's seems to be not working with go-grpc >= 1.19 [3]

Furthermore, we're exposing the proper GRPC server (the native go-grpc
one) on a different port. If decide to (finally!) expose our GRPC API,
that's the thing we should expose.

So, since we don't need this multiplexing business, we shouldn't keep
the code around. This is the cleanup.

[1]: https://github.com/philips/grpc-gateway-example/
[2]: https://github.com/grpc/grpc-go/issues/586
[3]: https://github.com/soheilhy/cmux/issues/64#issuecomment-491999193

Signed-off-by: Stephan Renatus <srenatus@chef.io>