WIP: Generate self-signed CA and separate certificate
This is a potential fix for reports of Chrome reporting
ERR_CERT_INVALID. Our suspicion is that Chrome may be rejecting the
use of a certificate with both `CA: TRUE` and the serverAuth extended
key usage.

Here we generated a self-signed CA and then generate a cert signed by
that CA.  Both certificates are concatenated for use in nginx.

Signed-off-by: Steven Danna <steve@chef.io>