Client Runs Narrowing Suggestions (#391)
Add Client runs search bar suggestion narrowing. The suggestions that appear in the search bar should be filtered to only show values from the nodes currently visible on the page. This filters the suggestions with the filters already selected in the search bar.

One exception is that the suggestions are not narrowed by the type they are for. For example, if a user is trying to add a new node name filter in the search bar the priorly added node name filters will not be used to narrow the values returned. This is because filters of the same type are or'ed together. Therefore the filters of the same type should not limit new filters.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>