Client runs regex filter as or (#401)
Updating the Client Runs search bar wildcard filtering to be an 'or' instead of an 'and'. Currently when matching with wildcards that term had to match all returned values. The desired behavior is if a second wildcard or exact term is created it matches should also be added to the returned results. This means that wildcard term should be combined with other matching types with an 'or'.
Signed-off-by: Lance Finfrock <lfinfrock@chef.io>