Revise project introspection rego to use only subjects
We can safely ignore action/resource pairs when
trying to answer the question of what projects a
user has access to.
The algorithm does not apply the same allow/deny rules.
Denied statements can simply be ignored, because access
to any bit of a project by some statement is sufficient to
say a user can access a given project.

Signed-off-by: michael sorens <msorens@users.noreply.github.com>