Provide unfiltering projects mechanism in UI
In order to populate the global projects filter,
need to be able to make a call to ListProjects without the
global projects filter itself being applied,
as it is on every call from the UI before this change.
This introduces the URL meta-parameter 'unfiltered'
so that any caller may make an arbitrary call unfiltered.

Signed-off-by: michael sorens <msorens@users.noreply.github.com>