Tailor introspect_projects to return needed data
1. Turn off project filtering via 'unfiltered' HTTP param.
Otherwise, introspect_projects would restrict its universe of discourse
to just the currently selected projects, when it should
always be looking at all allowed projects.
2. Allow unassigned_project to be returned.
We never want it to appear on the List Projects view,
but it should appear in the global projects filter
(when it is, in fact, an allowed project for the user).

One anomaly: Had to adjust the action on IntrospectAllProjects
in the proto file to *not* be under iam:introspect:*.
Without that change IntrospectAllProjects returns *all* projects rather than all *allowed* projects (because the system policy for iam:introspect:* includes projects: *).

Signed-off-by: michael sorens <msorens@users.noreply.github.com>