Rename auth get endpoints to list (#425)
Per my spreadsheet of auth endpoints (https://docs.google.com/spreadsheets/d/190c3VvCQGpsJ3eWq8DvqgTsIdZw3Mmk1os_9itjADgY/edit#gid=0) an anomaly stood out for GetTeams and GetUsers, which should be more appropriately ListTeams and ListUsers in v2.

This is done with this PR; also changed the interservice API for teams, as it seemed useful to do that, but we seem to only have a single interservice proto file for users. So not sure if it was worth adding a new v2 proto file just for that. Opinions?

* Change GetTeams to ListTeams in gateway
* Change GetTeams to ListTeams in teams-service
* Change GetUsers to ListUsers in gateway

Signed-off-by: michael sorens <msorens@users.noreply.github.com>