authz-service/server/v2: make IAM v2/v2.1 version switch a policy change event
Since there's two different stores for these now, it matters in a
multi-node setting which version the server is operating under. Since
the version is derived from the migration status, which in turn is
stored in the database, we make a change to that a "policy change
event": any instance that is connected to the database, but hasn't been
the one processing the migration GRPC call, will be informed that it
needs to update its store.

Figuring out which store to update is done by querying the database for
its migration status; it no longer is retrieved from local state of the
server instance.

Note: this does not notify other nodes in unhappy paths, since a failure
to migrate v1 policies should prohibit a switch to v2 (or v2.1).

Signed-off-by: Stephan Renatus <srenatus@chef.io>