Benchmarks for Applications Service (#495)
Currently, in the EAS Team, we are struggling a little bit to make decisions
in the backend, that is inside the `applications-service`.

We don't have a way to know the implications of adding or modifying code
and therefore, as engineers, we make lots of theories and deductions of
what "it could happen" based on assumptions.

An example is the following PR that is adding some extra complexity in
the query side of the micro-service. https://github.com/chef/automate/pull/490

This work is the first iteration to introduce benchmarks in our integration
tests to measure two main GRPC functions that we expose to consumers
(endpoints and the UI). These measurements will be recorded in our pipeline
so that we can understand how the system progress or degrades on code
changes.

# Definition of Done
After this change is merged, we will start recording times based on
Go Benchmarks for the GRPC functions:

- `GetServiceGroups` - Lists all service-groups in the system.
- `GetServices` - Lists all services in the system.

Here is the [(link)](https://buildkite.com/chef/chef-automate-master-verify-private/builds/1725#04637108-dcfb-4052-817e-1f8953d9f206) of the BuildKite Job that ran them

Output Gist: https://gist.github.com/afiune/f298f71788294d698e134dd82ad4b39c#file-pr-495-md

Signed-off-by: Salim Afiune <afiune@chef.io>