refresher/engine: add a rule cache to opa state
originally, i had a separate opa store for rules. but unmarshalling
the rules out of opa's non-typed store was too high a cost considering
we weren't even utilizing any of opa's capabilities to manipulate
the returned rules.

now, on authz-server init, authz gets the list of rules and loads them
into opa's rule cache. even though it's not technically an opa store,
i decided to keep the cache a part of opa since all of the HA wiring
is already in place for the opa cache.

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>