authz-service/{datamigrations,migrations}: add Makefile for timestamp versions
This is @phiggins idea -- it's _very_ unlikely that two PRs ever touch
the same migration again.

To create a new one, use `make new`. You can also pass a name:

    $ cd components/authz-service/storage/postgres/migration/sql/
    $ make new
    => 20190607130445_TODO.up.sql
    $ make new name=do_foo_things
    => 20190607130452_do_foo_things.up.sql

Of course, it's possible that the migrations of two different PRs do not
commute (i.e., A;B is different from B;A), but that situation has always
been a problem.

Signed-off-by: Stephan Renatus <srenatus@chef.io>