Migration node-state index to allow projects to be indexed (#476)
This change performs a migration on the node-state index to allow the projects field to be indexed. 

Added change of when a migration or elasticsearch mapping update fails the ingest-service will fail to start. 

Add case insensitive wildcard filters to the Client Runs search bar. This change adds a normalizer to the node-state fields that are being filtered on. This normalizer indexes the values of the field in lowercase. When filters for these fields are used the normalizer filters with lowercase values.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>