Use binlink-dir in a1 migration nightly tests
We install the latest habitat in that test. These changes should fix up
compatibility issues habitat-sh/habitat#6597, which reverts the behavior
of `hab pkg install -b` to what it used to be. For future compatibility,
I've used the `--binlink-dir` flag, which seems to be the new way of
doing things

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>