[A2-809] back up ListProjectRules API with real data (#517)
* refresher/engine: add a rule cache to opa state

originally, i had a separate opa store for rules. but unmarshalling
the rules out of opa's non-typed store was too high a cost considering
we weren't even utilizing any of opa's capabilities to manipulate
the returned rules.

now, on authz-server init, authz gets the list of rules and loads them
into opa's rule cache. even though it's not technically an opa store,
i decided to keep the cache a part of opa since all of the HA wiring
is already in place for the opa cache.

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz/test: update test files

with new setv2p1policies function definition

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz: update store on create/delete/update rule

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* (wip) authz: update ListProjectRules handler

to return correct response

this works as it expected on the command line. but the system integration
test setup needs to be updated so the test engine can be loaded with rules

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz: fix test setup to use real ruleRetriever

keep picturing a golden retriever when i see that :heart:

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz: drop testing logs

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz: drop no longer needed POC code

- don't need GetProjectRules anymore at all. it is now implemented as
ListRulesForProject(project_id) on the Projects API
- don't need any rule mapping rego anymore
- don't need mock rule structs

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* opa: drop benchmark

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz: update opa bindata

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz/opa: separate SetRules from V2P1SetPolicies

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz-service/storage/v2/memstore: bump policy version for rules changes

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* authz-service/storage/v2/postgres: bump policy version for rules changes

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* system_integration_test: unflake TestListProjectRules

It seems `assert.EqualValues` is stable on maps; `assert.Equal` fails
because map iteration order isn't deterministic.

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* conformance_v2_tests: SetRules on engine for v2.1

Signed-off-by: Stephan Renatus <srenatus@chef.io>

* authz: ListProjectRules -> ListRulesForAllProjects

to differentiate from the API ListRulesForProject(id)

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* authz/compliance/ingest: rename

ListProjectRules -> ListRulesForAllProject

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* address review comments

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* address review comments

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>

* api: update protobufs post-rebase

Signed-off-by: Brenna Hewer-Darroch <brenna@chef.io>